[configuration] Move repos package and "the" singleton to core
