+++ core/repos/repomanager.go
