+++ core/task/manager.go
@@ -36,8 +36,8 @@ import (
 	"github.com/AliceO2Group/Control/common/gera"
 	"github.com/AliceO2Group/Control/common/utils"
 	"github.com/AliceO2Group/Control/common/utils/uid"
-	"github.com/AliceO2Group/Control/configuration/repos"
-	"github.com/AliceO2Group/Control/configuration/the"
+	pb "github.com/AliceO2Group/Control/core/protos"
+	"github.com/AliceO2Group/Control/core/repos"
 	"github.com/AliceO2Group/Control/core/task/taskop"
 	"github.com/mesos/mesos-go/api/v1/lib/extras/store"
 	"gopkg.in/yaml.v3"
@@ -77,11 +77,13 @@ type Manager struct {
 	tasksFinished      int
 
 	schedulerState     *schedulerState
+	publicEventCh      chan<- *pb.Event
 	internalEventCh    chan<- event.Event
-	ackKilledTasks     *safeAcks
+	ackKilledTasks     map[string]chan struct{}
 }
 
 func NewManager(shutdown func(),
+	publicEventCh chan<- *pb.Event,
 	internalEventCh chan<- event.Event) (taskman *Manager, err error) {
 	// TODO(jdef) how to track/handle timeout errors that occur for SUBSCRIBE calls? we should
 	// probably tolerate X number of subsequent subscribe failures before bailing. we'll need
@@ -110,6 +112,7 @@ func NewManager(shutdown func(),
 	taskman = &Manager{
 		classes:            newClasses(),
 		roster:             newRoster(),
+		publicEventCh:      publicEventCh,
 		internalEventCh:    internalEventCh,
 	}
 	schedulerState, err := NewScheduler(taskman, fidStore, shutdown)
@@ -121,7 +124,7 @@ func NewManager(shutdown func(),
 	taskman.resourceOffersDone = taskman.schedulerState.resourceOffersDone
 	taskman.tasksToDeploy = taskman.schedulerState.tasksToDeploy
 	taskman.reviveOffersTrg = taskman.schedulerState.reviveOffersTrg
-	taskman.ackKilledTasks = newAcks()
+	taskman.ackKilledTasks = make(map[string]chan struct{})
 
 	schedulerState.setupCli()
 
@@ -494,6 +497,7 @@ func (m *Manager) configureTasks(envId uid.ID, tasks Tasks) error {
 	if len(strings.TrimSpace(errText)) != 0 {
 		return errors.New(response.Err().Error())
 	}
+
 	// FIXME: improve error handling ↑
 
 	return nil
@@ -620,7 +624,6 @@ func (m *Manager) updateTaskState(taskId string, state string) {
 	st := StateFromString(state)
 	taskPtr.state = st
 	taskPtr.safeToStop = false
-	taskPtr.SendEvent(&event.TaskEvent{Name: taskPtr.GetName(), TaskID: taskId, State: state, Hostname: taskPtr.hostname , ClassName: taskPtr.GetClassName()})
 	if taskPtr.GetParent() != nil {
 		taskPtr.GetParent().UpdateState(st)
 	}
@@ -634,7 +637,7 @@ func (m *Manager) updateTaskStatus(status *mesos.TaskStatus) {
 		log.WithField("taskId", taskId).
 			Warn("attempted status update of task not in roster")
 
-		if val, ok := m.ackKilledTasks.getValue(taskId); ok {
+		if val, ok := m.ackKilledTasks[taskId]; ok {
 			val <- struct{}{}
 		}
 
@@ -656,7 +659,6 @@ func (m *Manager) updateTaskStatus(status *mesos.TaskStatus) {
 			taskPtr.GetParent().UpdateStatus(INACTIVE)
 		}
 	}
-	taskPtr.SendEvent(&event.TaskEvent{Name: taskPtr.GetName(), TaskID: taskId, Status: taskPtr.status.String(), Hostname: taskPtr.hostname , ClassName: taskPtr.GetClassName()})
 }
 
 // Kill all tasks outside an environment (all unlocked tasks)
@@ -673,6 +675,7 @@ func (m *Manager) Cleanup() (killed Tasks, running Tasks, err error) {
 // Kill a specific list of tasks.
 // If the task list includes locked tasks, TaskNotFoundError is returned.
 func (m *Manager) KillTasks(taskIds []string) (killed Tasks, running Tasks, err error) {
+
 	toKill := m.roster.filtered(func(t *Task) bool {
 		if t.IsLocked() {
 			return false
@@ -691,15 +694,14 @@ func (m *Manager) KillTasks(taskIds []string) (killed Tasks, running Tasks, err 
 	}
 
 	for _, id := range toKill.GetTaskIds() {
-		m.ackKilledTasks.addAckChannel(id)
+		m.ackKilledTasks[id] = make(chan struct{})
 	}
 
 	killed, running, err = m.doKillTasks(toKill)
 	for _, id := range killed.GetTaskIds() {
-		ack,_ := m.ackKilledTasks.getValue(id)
-		<- ack
-		close(ack)
-		m.ackKilledTasks.deleteKey(id)
+		<- m.ackKilledTasks[id]
+		close(m.ackKilledTasks[id])
+		delete(m.ackKilledTasks, id)
 	}
 	return
 }
@@ -829,6 +831,8 @@ func (m *Manager) handleMessage(tm *TaskmanMessage) (error) {
 		}
 	case taskop.TaskStateMessage:
 		go m.updateTaskState(tm.taskId, tm.state)
+	// case event.KillTasks:
+		// m.killTasks(tm.GetTaskIds())
 	case taskop.ReleaseTasks:
 		go m.releaseTasks(tm.GetEnvironmentId(), tm.GetTasks())
 	}
