+++ core/task/task.go
@@ -46,9 +46,9 @@ import (
 	"github.com/AliceO2Group/Control/common/event"
 	"github.com/AliceO2Group/Control/common/utils/uid"
 	"github.com/AliceO2Group/Control/configuration/template"
-	"github.com/AliceO2Group/Control/configuration/the"
 	"github.com/AliceO2Group/Control/core/controlcommands"
 	"github.com/AliceO2Group/Control/core/task/channel"
+	"github.com/AliceO2Group/Control/core/the"
 	"github.com/mesos/mesos-go/api/v1/lib"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
