+++ core/workflow/aggregatorrole.go
@@ -29,10 +29,8 @@ import (
 	"strings"
 	texttemplate "text/template"
 
-	"github.com/AliceO2Group/Control/common/event"
-	"github.com/AliceO2Group/Control/configuration/repos"
 	"github.com/AliceO2Group/Control/configuration/template"
-	"github.com/AliceO2Group/Control/configuration/the"
+	"github.com/AliceO2Group/Control/core/repos"
 	"github.com/AliceO2Group/Control/core/task"
 	"github.com/gobwas/glob"
 	"github.com/sirupsen/logrus"
@@ -187,7 +185,6 @@ func (r *aggregatorRole) updateStatus(s task.Status) {
 		Debug("aggregator role about to merge incoming child status")
 	r.status.merge(s, r)
 	log.WithField("new status", r.status.get()).Debug("status merged")
-	r.SendEvent(&event.RoleEvent{Name: r.Name, Status: r.status.get().String(), RolePath: r.GetPath()})
 	if r.parent != nil {
 		r.parent.updateStatus(r.status.get())
 	}
@@ -199,7 +196,6 @@ func (r *aggregatorRole) updateState(s task.State) {
 	}
 	log.WithField("role", r.Name).WithField("state", s.String()).Debug("updating state")
 	r.state.merge(s, r)
-	r.SendEvent(&event.RoleEvent{Name: r.Name, State: r.state.get().String(), RolePath: r.GetPath()})
 	if r.parent != nil {
 		r.parent.updateState(r.state.get())
 	}
