+++ core/workflow/taskrole.go
@@ -30,10 +30,8 @@ import (
 	texttemplate "text/template"
 	"time"
 
-	"github.com/AliceO2Group/Control/common/event"
-	"github.com/AliceO2Group/Control/configuration/repos"
 	"github.com/AliceO2Group/Control/configuration/template"
-	"github.com/AliceO2Group/Control/configuration/the"
+	"github.com/AliceO2Group/Control/core/repos"
 	"github.com/AliceO2Group/Control/core/task"
 	"github.com/gobwas/glob"
 )
@@ -178,7 +176,6 @@ func (t *taskRole) updateStatus(s task.Status) {
 		log.WithField("status", s.String()).Error("cannot update status with nil parent")
 	}
 	t.status.merge(s, t)
-	t.SendEvent(&event.RoleEvent{Name: t.Name, Status: t.status.get().String(), RolePath: t.GetPath()})
 	t.parent.updateStatus(s)
 }
 
@@ -188,7 +185,6 @@ func (t *taskRole) updateState(s task.State) {
 	}
 	log.WithField("role", t.Name).WithField("state", s.String()).Debug("updating state")
 	t.state.merge(s, t)
-	t.SendEvent(&event.RoleEvent{Name: t.Name, State: t.state.get().String(), RolePath: t.GetPath()})
 	t.parent.updateState(s)
 }
